PR comments: moved functions to vcftools.py and added tests
