+++ test/system/test_correct_genotypes_by_imputation.py
@@ -4,15 +4,43 @@ import subprocess
 from test import get_resource_dir, test_dir
 
 import pandas as pd
+import numpy as np
 
-from ugvc.pipelines import correct_genotypes_by_imputation
-
+from ugvc.pipelines import correct_genotypes_by_imputation as corgen
+from ugvc.vcfbed.vcftools import genotype_ordering
 
 class TestCorrectGenotypesByImputation:
     inputs_dir = get_resource_dir(__file__)
     resources_dir = get_resource_dir(__file__)
 
-    def test_correct_genotypes_by_imputation(self, tmpdir):
+    def test_convert_ds_to_genotype_imputation_priors(self):
+        # Homozygous biallelic variant
+        ds_ar = np.array([2])
+        gt_ar = genotype_ordering(1)
+        num_alt = 1
+        epsilon = 0.01
+        assert np.all(corgen._convert_ds_to_genotype_imputation_priors(ds_ar, gt_ar, num_alt, epsilon) == np.array([1,0.01,0.99]))
+        # Heterozygous biallelic variant
+        ds_ar = np.array([1])
+        gt_ar = genotype_ordering(1)
+        num_alt = 1
+        epsilon = 0.01
+        assert np.all(corgen._convert_ds_to_genotype_imputation_priors(ds_ar, gt_ar, num_alt, epsilon) == np.array([1,0.99,0.01]))
+        # Heterozygous tri-allelic variant
+        ds_ar = np.array([1,1])
+        gt_ar = genotype_ordering(2)
+        num_alt = 2
+        epsilon = 0.01
+        assert np.all(corgen._convert_ds_to_genotype_imputation_priors(ds_ar, gt_ar, num_alt, epsilon) == np.array([1,0.99,0.01,0.99,0.99,0.01]))
+        # Heterozygous tri-allelic variant with missing imputation data
+        ds_ar = np.array([2,None],dtype=float)
+        gt_ar = genotype_ordering(2)
+        num_alt = 2
+        epsilon = 0.01
+        assert np.all(corgen._convert_ds_to_genotype_imputation_priors(ds_ar, gt_ar, num_alt, epsilon) == np.array([1,0.01,0.99,0.01,0.01,0.01]))
+
+
+    def test_corgen(self, tmpdir):
         os.makedirs(tmpdir, exist_ok=True)
         with open(os.path.join(tmpdir, "chrom_to_cohort_vcf.json"), "w") as f:
             json.dump(
@@ -32,7 +60,7 @@ class TestCorrectGenotypesByImputation:
             )
         input_vcf = os.path.join(self.resources_dir, "dv_013783-X0013.vcf.gz")
         output_vcf = os.path.join(tmpdir, "output.vcf.gz")
-        correct_genotypes_by_imputation.run(
+        corgen.run(
             [
                 "correct_genotypes_by_imputation",
                 "--input_vcf",
