+++ test/unit/vcfbed/test_vcftools.py
@@ -2,6 +2,7 @@ from os.path import join as pjoin
 from test import get_resource_dir
 
 import pandas as pd
+import numpy as np
 
 import ugvc.vcfbed.vcftools as vcftools
 
@@ -189,3 +190,27 @@ class TestGetVcfDf:
         ]
         for x in ignore_fields:
             assert x not in df.columns
+
+
+def test_phred():
+    assert np.all(vcftools.phred((0.1,0.01,0.001))==np.array([10., 20., 30.]))
+
+def test_unphred():
+    assert np.all(vcftools.unphred((10,20,30))==np.array([0.1, 0.01, 0.001]))
+
+def test_different_gt():
+    assert not vcftools.different_gt('0|1', (0, 1))
+    assert not vcftools.different_gt('1|0', (0, 1))
+    assert vcftools.different_gt('1|1', (0, 1))
+    assert vcftools.different_gt('1/2', (0, 1))
+
+# @pytest.mark.parametrize("extra_params,expected", zip([], expected_outputs))
+# def test_phred( extra_params, expected):
+#
+#     input_file = pjoin(get_resource_dir(__file__), "120461-BC23.for_test_input.bam")
+#     shutil.copyfile(input_file, pjoin(tmpdir, basename(input_file)))
+#     cmd = [trimming_script, pjoin(tmpdir, basename(input_file))] + extra_params
+#     subprocess.check_call(cmd, cwd=tmpdir)
+#     output_file = pjoin(tmpdir, basename(input_file).replace("bam", "with_adapter_tags.bam"))
+#     assert exists(output_file)
+#     assert _compare_bam_records(output_file, expected), f"{output_file} and {expected} are not identical"
